+++ src/views.rs
@@ -1,2 +1,4 @@
 pub mod block_group;
+pub mod block_group_viewer;
+pub mod block_layout;
 pub mod patch;
