+++ src/views/block_group.rs
@@ -3,274 +3,37 @@ use crate::models::{
     edge::Edge,
     sample::Sample,
 };
+use crate::views::block_group_viewer::{Viewer, PlotParameters, ScrollState};
+use crate::views::block_layout::ScaledLayout;
 
-use crossterm::event::KeyEventKind;
-use itertools::Itertools; use noodles::vcf::header::record::value::map::info;
-// for tuple_windows
 use rusqlite::Connection;
 
 use core::panic;
 use std::collections::{HashMap, HashSet};
 use std::error::Error;
-use std::io::Write;
 use std::time::{Duration, Instant};
 use std::u32;
 
 use log::info;
 
-
 use crossterm::{
-    event::{self, KeyCode, KeyModifiers},
+    event::{self, KeyCode, KeyModifiers, KeyEventKind},
     execute,
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
 };
 use ratatui::{
     init,
     backend::{Backend, CrosstermBackend},
-    layout::Rect,
-    widgets::{Block, Borders, Paragraph, Wrap},
+    layout::{Rect, Constraint},
+    widgets::{Block, Borders, Paragraph, Wrap, Clear},
     widgets::canvas::{Canvas, Line},
     Terminal,
-    style::{Style, Color},
+    style::{Style, Color, Stylize},
     text::{Span, Text},
 };
 use rust_sugiyama::{configure::Config, from_edges};
 
-/// Holds current scrolling offset and a zoom factors for data units per terminal cell.
-///
-/// - `scale` = data units per 1 terminal cell.  
-///   - If `scale` = 1.0, each cell is 1 data unit.  
-///   - If `scale` = 2.0, each cell is 2 data units (you see *more* data).  
-///   - If `scale` = 0.5, each cell is 0.5 data units (you see *less* data, zoomed in).
-/// - `aspect_ratio` = width / height of a terminal cell in data units.
-/// - `block_len` = how much of the sequence to show in each block label.
-/// 
-struct ScrollState {
-    offset_x: i32,
-    offset_y: i32
-}
 
-/// Holds data for the viewer.
-struct Viewer {
-    layout: ScaledLayout, // Coordinates and labels
-    scroll: ScrollState,
-    plot_area: Rect, 
-    plot_parameters: PlotParameters,
-}
-
-/// Holds the parameters for plotting the graph.
-struct PlotParameters {
-    label_width: u32, // Truncate each label to this width
-    scaling: u32, // Stretch the edges by this factor, taking into account the aspect ratio
-    aspect_ratio: f32, // Scale differently in x and y directions
-}
-
-/// Holds processed and scaled layout data.
-/// - `lines` = pairs of coordinates for each edge.
-/// - `labels` = truncated sequences or symbols for each block.
-/// - `highlight_[a|b]` = block ID or (block ID, coordinate) to highlight in color A or B.
-/// The raw layout from the Sugiyama algorithm is processed as follow:
-/// - The coordinates are rounded to the nearest integer and transposed to go from top-to-bottom to left-to-right.
-/// - Each block is assigned a layer (or rank) based on its y-coordinate.
-/// - The width of each layer is determined by the widest label in that layer.
-/// - The distance between layers is scaled horizontally and vertically 
-struct ScaledLayout {
-    lines: Vec<((f64, f64), (f64, f64))>, // Pairs of coordinates for each edge
-    labels: HashMap<u32, (String, u32, u32)>, // K:Block ID, V: label, x, y
-    highlight_a: Option<(u32, Option<(u32, u32)>)>, // Block ID or (Block ID, coordinate) to highlight in color A
-    highlight_b: Option<(u32, Option<(u32, u32)>)>, // Block ID or (Block ID, coordinate) to highlight in color B
-    // Input data that is stored but not supposed to be used directly
-    #[doc(hidden)]
-    _edges: Vec<(u32, u32)>, // Block ID pairs
-    _raw_layout: Vec<(u32, (f64, f64))>, // Raw layout from the Sugiyama algorithm
-    _sequences: Option<HashMap<u32, String>>, // Block ID to sequence (full length)
-
-}
-
-impl ScaledLayout {
-    fn new(
-        raw_layout: Vec<(u32, (f64, f64))>, // Block ID, (x, y) coordinates
-        edges: Vec<(u32, u32)>, // Block ID pairs
-        parameters: &PlotParameters,
-        sequences: Option<HashMap<u32, String>>
-    ) -> Self {
-        let mut layout = ScaledLayout {
-            lines: Vec::new(),
-            labels: HashMap::new(),
-            highlight_a: None,
-            highlight_b: None,
-            _raw_layout: raw_layout,
-            _edges: edges,
-            _sequences: sequences
-        };
-        layout.rescale(parameters);
-        layout
-    }
-    pub fn rescale(&mut self, parameters: &PlotParameters) {
-        // Scale the overall layout, round it to the nearest integer, transpose x and y, and sort by x-coordinate
-        let scale_x = parameters.scaling as f64;
-        let scale_y = parameters.scaling as f64 * parameters.aspect_ratio as f64;
-        let layout: Vec<(u32, (u32, u32))> = self._raw_layout.iter()
-            .map(|(id, (x, y))| (*id, ((y * scale_x).round() as u32, (x * scale_y).round() as u32)))
-            .sorted_by(|a, b| a.1 .0.cmp(&b.1 .0))
-            .collect();
-
-        // We can stop here if:
-        // - the target label width is < 5 or no labels were given
-        if self._sequences.is_none() || (parameters.label_width < 5)  {
-            // Turn the layout into a hashmap so we can easily look up the coordinates for each block
-            let layout: HashMap<u32, (u32, u32)> = layout.iter().map(|(id, (x, y))| (*id, (*x, *y))).collect();
-            self.lines = self._edges.iter()
-                .map(|(source, target)| {
-                    let source_coord = layout.get(source).map(|&(x, y)| (x as f64 + 0.5, y as f64 + 0.25)).unwrap();
-                    let target_coord = layout.get(target).map(|&(x, y)| (x as f64 - 1.0, y as f64 + 0.25)).unwrap();
-                    (source_coord, target_coord)
-                })
-                .collect();
-            self.labels = layout.iter()
-                .map(|(id, (x, y))| (*id, ("●".to_string(), *x, *y)))
-                .collect();
-            return;
-        }
-
-        // Loop over the sorted layout and group the blocks by rank (y-coordinate)
-        let mut processed_layout: Vec<(u32, String, (u32, u32))> = Vec::new(); //
-        let mut current_x = layout[0].1 .0;
-        let mut current_layer: Vec<(u32, String, u32)> = Vec::new(); // Block ID, label, y-coordinate
-        let mut layer_width = std::cmp::min(self._sequences.as_ref().unwrap().get(&layout[0].0).unwrap().len() as u32, parameters.label_width);
-        let mut cumulative_offset = 0;
-        for (id, (x, y)) in layout.iter() {
-            let full_label = self._sequences
-                .as_ref()
-                .and_then(|labels| labels.get(id))
-                .unwrap();
-            let truncated_label = inner_truncation(full_label, parameters.label_width);
-                    
-            if *x == current_x {
-                // This means we are still in the same layer
-                // Keep a tally of the maximum label width
-                layer_width = std::cmp::max(layer_width, truncated_label.len() as u32);
-
-                // Add the block to the current layer vector
-                current_layer.push((*id, truncated_label, *y));
-            } else {
-                // We switched to a new layer
-                // Loop over the current layer and:
-                // - increment the x-coordinate by the cumulative offset so far
-                // - horizontally center the block in its layer
-                for (id, label, y) in current_layer {
-                    let centering_offset = (layer_width - label.len() as u32) / 2;
-                    let x = current_x + centering_offset + cumulative_offset;
-                    // Store the new x-coordinate and truncated label in the combined vector
-                    processed_layout.push((id, label, (x, y)));
-                }
-                // Increment the cumulative offset for the next layer by the width of the current layer
-                cumulative_offset += layer_width;
-
-                // Reset the layer width and the current layer
-                layer_width = truncated_label.len() as u32;
-                current_layer = vec![(*id, truncated_label, *y)];
-                current_x = *x;
-            }
-        }
-        // Loop over the last layer (wasn't processed yet)
-        for (id, label, y) in current_layer {
-            let centering_offset = (layer_width - label.len() as u32) / 2;
-            let x = current_x + centering_offset + cumulative_offset;
-            processed_layout.push((id, label, (x, y)));
-        }
-
-        // Make a hashmap of the processed layout so we can quickly find labels with coordinates
-        self.labels = processed_layout.into_iter().map(|(id, label, (x, y))| (id, (label, x, y))).collect();
-
-        // Recalculate all the edges so they meet labels on the sides instead of the center
-        self.lines = self._edges.iter()
-            .map(|(source, target)| {
-            let (source_label, source_x, source_y) = self.labels.get(source).unwrap();
-            let (_, target_x, target_y) = self.labels.get(target).unwrap();
-            let source_x = *source_x as f64 + source_label.len() as f64;
-            let source_y = *source_y as f64 + 0.5;
-            let target_x = *target_x as f64 - 1.5;
-            let target_y = *target_y as f64 + 0.5;
-            ((source_x, source_y), (target_x, target_y))
-            })
-            .collect();
-
-    }
-}
-
-/// Truncate a string to a certain length, adding an ellipsis in the middle
-fn inner_truncation(s: &str, target_length: u32) -> String {
-    let input_length = s.len() as u32;
-    if input_length <= target_length {
-        return s.to_string();
-    } else if target_length < 3 {
-        return "●".to_string();
-    }
-    // length - 3 because we need space for the ellipsis
-    let left_len = (target_length-3) / 2 + ((target_length-3)  % 2);
-    let right_len = (target_length-3) - left_len;
-    
-    format!("{}...{}", &s[..left_len as usize], 
-        &s[input_length as usize - right_len as usize..])
-}
-
-
-/// Draw the canvas, ensuring a 1 to 1 mapping between data units and terminal cells.
-fn draw_scrollable_canvas(frame: &mut ratatui::Frame, viewer: &Viewer) {
-    // Set up the coordinate systems for the window and the canvas
-
-    // Terminal window coordinates
-    let plot_area = viewer.plot_area;
-    let block = Block::default().borders(Borders::ALL).title("graph viewer");
-    let usable_area = block.inner(plot_area);
-
-    // Data coordinates (the top-left corner of our view is (offset_x, offset_y))
-    let viewport_left = viewer.scroll.offset_x;
-    let viewport_right = viewer.scroll.offset_x + usable_area.width as i32;
-    let viewport_top = viewer.scroll.offset_y;
-    let viewport_bottom = viewer.scroll.offset_y + usable_area.height as i32;
-
-    // Create the canvas
-    let canvas = Canvas::default()
-        .block(block)
-        // Adjust the x_bounds and y_bounds by the scroll offsets.
-        .x_bounds([viewport_left as f64, viewport_right as f64])
-        .y_bounds([viewport_top as f64, viewport_bottom as f64])
-        .paint(|ctx| {
-            // Draw the lines described in the processed layout
-            for ((x1, y1), (x2, y2)) in &viewer.layout.lines {
-                // Clip the line to the visible area, skip if it's not visible itself
-                if let Some(((x1c, y1c), (x2c, y2c))) = clip_line((*x1, *y1), (*x2, *y2), 
-                    (viewport_left as f64, viewport_top as f64), 
-                    (viewport_right as f64, viewport_bottom as f64)) {
-                    ctx.draw(&Line {
-                        x1: x1c,
-                        y1: y1c,
-                        x2: x2c,
-                        y2: y2c,
-                        color: Color::White,
-                    });
-                }
-            }
-            // Print the labels
-            for (block_id, (label, x, y)) in &viewer.layout.labels {
-                // Skip labels that are not in the visible area (vertical)
-                // (note that the z-axis is upside down)
-                if (*y as i32) < viewport_top || (*y as i32) > viewport_bottom {
-                    continue;
-                }
-                // Clip labels that are potentially in the window (horizontal)
-                let clipped_label = clip_label(label, *x as isize, 
-                    viewport_left as isize, usable_area.width as usize);
-                if !clipped_label.is_empty() {
-                    ctx.print((*x as isize).max(viewport_left as isize) as f64, *y as f64, clipped_label.clone());
-                }
-            }
-
-        });
-    frame.render_widget(canvas, plot_area);   
-}
 
 pub fn view_block_group(
     conn: &Connection,
@@ -351,6 +114,9 @@ pub fn view_block_group(
         assert!(layout.iter().any(|(id, _)| *id == block.id as u32), "Block ID {} not found in layout", block.id);
     }
 
+
+    // TODO: the below code should go into its own module
+
     // Setup terminal
     enable_raw_mode()?;
     let mut stdout = std::io::stdout();
@@ -359,12 +125,9 @@ pub fn view_block_group(
 
 
     // Initialize viewer state
-    // TODO: set up the viewer with sensible defaults and add parameters later
-    // or borrow parameters instead
-    // 
     let initial_parameters = PlotParameters {
         label_width: 11,
-        scaling: 2,
+        scale: 2,
         aspect_ratio: 0.5
     };
 
@@ -378,14 +141,15 @@ pub fn view_block_group(
     let scaled_layout = ScaledLayout::new(layout, block_pairs_u32, &initial_parameters, Some(sequences));
 
     // Calculate the center point of the coordinates from the labels in the ScaledLayout
-    let center_x = scaled_layout.labels.values().map(|(_, x, _)| x).sum::<u32>() as f64 / scaled_layout.labels.len() as f64;
+    //let center_x = scaled_layout.labels.values().map(|(_, x, _)| x).sum::<u32>() as f64 / scaled_layout.labels.len() as f64;
     let center_y = scaled_layout.labels.values().map(|(_, _, y)| y).sum::<u32>() as f64 / scaled_layout.labels.len() as f64;
 
-    // Set the initial offset so that the center point is in the center of the terminal
+    // Set the initial offset so that graph is vertically centered, and left-aligned with some margin
     // We haven't set up a canvas yet, so we don't know more about the plot area
     let initial_scroll_state = ScrollState {
-        offset_x: center_x.round() as i32 - (terminal.get_frame().area().width as f64 / 2.0).round() as i32,
+        offset_x: -5,
         offset_y: center_y.round() as i32 - (terminal.get_frame().area().height as f64 / 2.0).round() as i32,
+        selected_block: None,
     };
 
     let mut viewer = Viewer {
@@ -394,7 +158,7 @@ pub fn view_block_group(
         plot_area: Rect::default(),
         plot_parameters: PlotParameters {
             label_width: 11,
-            scaling: 2,
+            scale: 2,
             aspect_ratio: 0.5
         },
     };
@@ -402,31 +166,84 @@ pub fn view_block_group(
     // Basic event loop
     let tick_rate = Duration::from_millis(100);
     let mut last_tick = Instant::now();
+    let mut show_panel = false; // Informational popup
     loop {
         // Draw the UI
         terminal.draw(|frame| {
-            // Currently we have just 2 widgets: the canvas and the status bar
-            // For anything more complex we should use a Layout to split the frame area
+            /// A layout consisting of a canvas and a status bar, with optionally a panel
+            /// - The canvas is where the graph is drawn
+            /// - The status bar is where the controls are displayed
+            /// - The panel is a scrollable paragraph that can be toggled on and off
+            
             let status_bar_height: u16 = 1;
-            let status_bar_area = Rect::new(frame.area().left(), 
-                                                frame.area().bottom()-status_bar_height, 
-                                                 frame.area().width, 
-                                                  status_bar_height);
 
-            // Canvas
-            let viewer = &mut viewer;
-            viewer.plot_area = Rect::new(frame.area().left(), frame.area().top(), 
-                                frame.area().width, frame.area().height-status_bar_height); 
-            draw_scrollable_canvas(frame, viewer);
+            // Define the layouts
+            // The outer layout is a vertical split between the canvas and the status bar
+            // The inner layout is a vertical split between the canvas and the panel
+
+            let outer_layout = ratatui::layout::Layout::default()
+                .direction(ratatui::layout::Direction::Vertical)
+                .constraints(vec!
+                    [
+                        ratatui::layout::Constraint::Min(1),
+                        ratatui::layout::Constraint::Length(status_bar_height),
+                    ]
+                )
+                .split(frame.area());
+
+            let inner_layout = ratatui::layout::Layout::default()
+                .direction(ratatui::layout::Direction::Vertical)
+                .constraints(vec!
+                    [
+                        Constraint::Percentage(75),
+                        Constraint::Percentage(25),
+                    ]
+                )
+                .split(outer_layout[0]);
+
+            
+            let canvas_area = if show_panel { inner_layout[0] } else { outer_layout[0] }; 
+            let panel_area = if show_panel { inner_layout[1] } else { Rect::default() };
+            let status_bar_area = outer_layout[1];
+
+            // Ask the viewer to paint the canvas
+            viewer.paint_canvas(frame, canvas_area);
 
             // Status bar
             let status_bar_contents = format!(
                 "{:width$}",
-                " arrows=scroll, +/-=zoom (hold shift to go faster), q=quit",
-                width = frame.area().width as usize);
+                " Controls: arrows(+shift)=scroll, +/-=zoom, tab=select blocks, return=show information on block, q=quit",
+                width = status_bar_area.width as usize);
+
             let status_bar = Paragraph::new(Text::styled(status_bar_contents, 
                 Style::default().bg(Color::DarkGray).fg(Color::White)));
+
             frame.render_widget(status_bar, status_bar_area);
+
+            // Panel
+            if show_panel {
+                let panel_block = Block::default().borders(Borders::ALL);
+                
+                let content_area = panel_block.inner(panel_area);
+                let mut panel_text = Text::from("No content found");
+
+                // Get information about the currently selected block
+                if viewer.scroll.selected_block.is_some() {
+                    let selected_block = viewer.scroll.selected_block.unwrap();
+                    let block = blocks.iter().find(|block| block.id == selected_block as i64).unwrap();
+                    panel_text = Text::from(format!("Block ID: {}\nNode ID: {}\nStart: {}\nEnd: {}\n", 
+                        block.node_id, block.id, block.start, block.end));
+                } 
+
+                let panel_content = Paragraph::new(panel_text)
+                    .wrap(Wrap { trim: true })
+                    .scroll((0, 0))
+                    .style(Style::default().bg(Color::Reset));
+  
+                // First clear the area, then render
+                frame.render_widget(Clear, content_area);
+                frame.render_widget(panel_content, content_area);
+            }
         })?;
 
         // Handle input
@@ -436,11 +253,12 @@ pub fn view_block_group(
         if crossterm::event::poll(timeout)? {
             if let event::Event::Key(key) = event::read()? {
                 if key.kind == KeyEventKind::Press {
+                    // Exit on q
+                    if key.code == KeyCode::Char('q') {
+                        break;
+                    }
                     match key.code {
-                        KeyCode::Char('q') => {
-                            // Exit on 'q'
-                            break;
-                        }
+                    // Scrolling through the graph
                         KeyCode::Left => {
                             // Scroll left
                             if key.modifiers == KeyModifiers::SHIFT {
@@ -448,6 +266,8 @@ pub fn view_block_group(
                             } else {
                                 viewer.scroll.offset_x -= 1;
                             }
+                            // Forget the selected block if it's not visible
+                            viewer.unselect_if_not_visible();
                         }
                         KeyCode::Right => {
                             // Scroll right
@@ -456,6 +276,7 @@ pub fn view_block_group(
                             } else {
                                 viewer.scroll.offset_x += 1;
                             }
+                            viewer.unselect_if_not_visible();
                         }
                         KeyCode::Up => {
                             // Scroll up
@@ -464,6 +285,7 @@ pub fn view_block_group(
                             } else {
                                 viewer.scroll.offset_y += 1;
                             }
+                            viewer.unselect_if_not_visible();
                         }
                         KeyCode::Down => {
                             // Scroll down
@@ -472,31 +294,41 @@ pub fn view_block_group(
                             } else {
                                 viewer.scroll.offset_y -= 1;
                             }
+                            viewer.unselect_if_not_visible();
                         }
+                    // Zooming in and out
                         KeyCode::Char('+') | KeyCode::Char('=') => {
                             // Increase how much of the sequence is shown in each block label: 0 vs 11 vs 100 characters.
                             // 11 was picked as the default because it results in symmetrical labels.
                             // After 100 it just becomes the full length.
-                            viewer.plot_parameters.label_width = match viewer.plot_parameters.label_width {
-                                0 => 11,
-                                11 => 100,
-                                100 => u32::MAX,
-                                _ => u32::MAX,
-                            };
-                            // Once we're at the full length, start increasing the scale instead
                             if viewer.plot_parameters.label_width == u32::MAX {
-                                viewer.plot_parameters.scaling += 1;
-                            }
+                                // If we're already maximizing the length, start increasing the scale
+                                viewer.plot_parameters.scale += 1;
+                            } else {
+                                // Otherwise, increase the label width
+                                viewer.plot_parameters.label_width = match viewer.plot_parameters.label_width {
+                                    0 => 11,
+                                    11 => 100,
+                                    100 => u32::MAX,
+                                    _ => u32::MAX,
+                                }
+                            };
+                            // Recalculate the layout
                             viewer.layout.rescale(&viewer.plot_parameters);
+
+                            // Center the viewport on the selected block if there is one
+                            if let Some(selected_block) = viewer.scroll.selected_block {
+                                viewer.center_on_block(selected_block);
+                            }
                         }
                         KeyCode::Char('-') | KeyCode::Char('_') => {
                             // Decrease how much of the sequence is shown in each block label: 11 vs 100 vs 1000 characters.
                             // 11 was picked as the default because it results in symmetrical labels.
                             // After 1000 it just becomes the full length.
 
-                            if viewer.plot_parameters.scaling > 1 {
+                            if viewer.plot_parameters.scale > 1 {
                                 // Decrease the scale if we're not at the minimum scale (1)
-                                viewer.plot_parameters.scaling -= 1;
+                                viewer.plot_parameters.scale -= 1;
                             } else {
                                 // If we're at the minimum scale, start decreasing the label width
                                 viewer.plot_parameters.label_width = match viewer.plot_parameters.label_width {
@@ -507,6 +339,35 @@ pub fn view_block_group(
                                 };
                             }
                             viewer.layout.rescale(&viewer.plot_parameters);
+                            if let Some(selected_block) = viewer.scroll.selected_block {
+                                viewer.center_on_block(selected_block);
+                            }
+                        }
+                    // Performing actions on blocks
+                        KeyCode::Tab => {
+                            // Cycle through visible blocks in the viewport
+                            viewer.cycle_blocks(false);
+                        }
+                        KeyCode::BackTab => {
+                            // Reverse cycle through visible blocks in the viewport
+                            viewer.cycle_blocks(true);
+                        }
+                        KeyCode::Esc => {
+                            // If we have a popup open, close it, otherwise unselect the selected block
+                            if show_panel {
+                                show_panel = false;
+                            } else {
+                                viewer.scroll.selected_block = None;
+                            }
+                        }    
+                        KeyCode::Enter => {
+                            // Show information on the selected block
+                            //viewer.show_block_info();
+                            if let Some(selected_block) = viewer.scroll.selected_block {
+                                show_panel = true;
+
+                            }
+                            
                         }
                         _ => {}
                     }
@@ -527,362 +388,10 @@ pub fn view_block_group(
     Ok(())
 }
 
-/// Clips a string to a specific window, indicating that it has been clipped.
-/// - If the string is empty, it returns an empty string.
-/// - If the string is shorter than the window, it returns the string.
-/// - If the string is longer than the window, it clips the string and replaces the last character with a period.
-/// - If the string is not within the window at all, it returns an empty string.
-pub fn clip_label(label: &str, label_start: isize, window_start: isize, window_width: usize) -> String {
-    if label.is_empty() {
-        return "".to_string();
-    }
-    let label_end = label_start + label.len() as isize - 1;
-    let window_end = window_start + window_width as isize - 1;
-    if label_end < window_start || label_start > window_end {
-        return "".to_string();
-    }
- 
-    if label_start >= window_start && label_end <= window_end {
-        return label.to_string();
-    }
 
-    let mut clipped = label.to_string();
-
-    // Process the right side first so we don't lose alignment:
-    let delta_right = label_end - window_end;
-    if delta_right > 0 {
-        clipped.replace_range((label.len() as isize - delta_right - 1) as usize.., "…");
-    }
-    let delta_left = window_start - label_start;
-    if delta_left > 0 {
-        clipped.replace_range(..delta_left as usize + 1, "…");
-    }
-
-    clipped
-}
-
-/// Clip a line given as two points to a specific window, also given by two points
-/// - If the line is completely outside the window, it returns None.
-/// - If the line is completely inside the window, it returns the original line.
-/// - If the line is partially inside the window, it clips the line to the window.
-/// 
-/// This may be made more efficient through bitwise comparisons (see Cohen-Sutherland line clipping algorithm)
-pub fn clip_line(
-    (x1, y1): (f64, f64),
-    (x2, y2): (f64, f64),
-    (wx1, wy1): (f64, f64),
-    (wx2, wy2): (f64, f64),
-) -> Option<((f64, f64), (f64, f64))> {
-    let mut t0 = 0.0;
-    let mut t1 = 1.0;
-    let dx = x2 - x1;
-    let dy = y2 - y1;
-
-    let clip = |p: f64, q: f64, t0: &mut f64, t1: &mut f64| -> bool {
-        if p == 0.0 {
-            return q >= 0.0;
-        }
-        let r = q / p;
-        if p < 0.0 {
-            if r > *t1 {
-                return false;
-            }
-            if r > *t0 {
-                *t0 = r;
-            }
-        } else {
-            if r < *t0 {
-                return false;
-            }
-            if r < *t1 {
-                *t1 = r;
-            }
-        }
-        true
-    };
-
-    if clip(-dx, x1 - wx1, &mut t0, &mut t1)
-        && clip(dx, wx2 - x1, &mut t0, &mut t1)
-        && clip(-dy, y1 - wy1, &mut t0, &mut t1)
-        && clip(dy, wy2 - y1, &mut t0, &mut t1)
-    {
-        let nx1 = x1 + t0 * dx;
-        let ny1 = y1 + t0 * dy;
-        let nx2 = x1 + t1 * dx;
-        let ny2 = y1 + t1 * dy;
-        Some(((nx1, ny1), (nx2, ny2)))
-    } else {
-        None
-    }
-}
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    fn test_clip_line_helper(
-        x1: f64,
-        y1: f64,
-        x2: f64,
-        y2: f64,
-        wx1: f64,
-        wy1: f64,
-        wx2: f64,
-        wy2: f64,
-        expected: Option<((f64, f64), (f64, f64))>,
-    ) {
-        let clipped = clip_line((x1, y1), (x2, y2), (wx1, wy1), (wx2, wy2));
-        assert_eq!(clipped, expected);
-    }
-
-    #[test]
-    fn test_clip_line_outside() {
-        test_clip_line_helper(0.0, 0.0,
-                              1.0, 1.0, 
-                              2.0, 2.0, 
-                              3.0, 3.0, 
-                              None);
-    }
-
-    #[test  ]
-    fn test_clip_line_inside() {
-        test_clip_line_helper(0.0, 0.0,
-                              1.0, 1.0, 
-                              -1.0, -1.0, 
-                              1.5, 1.5, 
-                              Some(((0.0, 0.0), (1.0, 1.0))));
-    }
-
-    #[test  ]
-    fn test_clip_line_partial() {
-        test_clip_line_helper(0.0, 0.0,
-                              1.0, 1.0, 
-                              0.5, 0.5, 
-                              1.5, 1.5, 
-                              Some(((0.5, 0.5), (1.0, 1.0))));
-    }
-
-    #[test]
-    fn test_clip_label_negative_offset() {
-        //  -2   0 1 2 3 4 5 6 7 8 9 
-        //  [        A B C]D E  
-        let clipped = clip_label("ABCDE", 2, -2, 7);
-        assert_eq!(clipped, "AB…");
-    }
-
-    #[test]
-    fn test_clip_label_internal() {
-        // 0 1 2 3 4 5 6 7 8 9 
-        //  [  A B C D E  ] 
-        let clipped = clip_label("ABCDE", 2, 1, 7);
-        assert_eq!(clipped, "ABCDE");
-    }
-
-    #[test]
-    fn test_clip_label_external() {
-        // 0 1 2 3 4 5 6 7 8 9 
-        //     A B  [     ] 
-        let clipped = clip_label("AB", 2, 5, 3);
-        assert_eq!(clipped, "");
-    }
-
-    #[test]
-    fn test_clip_label_left() {
-        // 0 1 2 3 4 5 6 7 8 9 
-        //     A B[C D E F G H] 
-        let clipped = clip_label("ABCDEFGH", 2, 4, 10);
-        assert_eq!(clipped, "…DEFGH");
-    }
-
-    #[test]
-    fn test_clip_label_right() {
-        // 0 1 2 3 4 5 6 7 8 9 
-        //[    A B C D E]F G H
-        let clipped = clip_label("ABCDEFGH", 2, 0, 7);
-        assert_eq!(clipped, "ABCD…");
-    }
-
-    #[test]
-    fn test_clip_label_both() {
-        // 0 1 2 3 4 5 6 7 8 9 
-        //     A B[C D E]F G H
-        let clipped = clip_label("ABCDEFGH", 2, 4, 3);
-        assert_eq!(clipped, "…D…");
-    }
-
-    #[test]
-    fn test_inner_truncation_no_truncation_needed() {
-        let s = "hello";
-        let truncated = inner_truncation(s, 5);
-        assert_eq!(truncated, "hello");
-    }
-
-    #[test]
-    fn test_inner_truncation_truncate_to_odd_length() {
-        let s = "hello world";
-        let truncated = inner_truncation(s, 5);
-        assert_eq!(truncated, "h...d");
-    }
-
-    #[test]
-    fn test_inner_truncation_truncate_to_even_length() {
-        let s = "hello world";
-        let truncated = inner_truncation(s, 6);
-        assert_eq!(truncated, "he...d");
-    }
-
-    #[test]
-    fn test_inner_truncation_empty_string() {
-        let s = "";
-        let truncated = inner_truncation(s, 5);
-        assert_eq!(truncated, "");
-    }
-
-    #[test]
-    fn test_scaled_layout_new_unlabeled() {
-        let _ = env_logger::try_init();
-
-        let edges = vec![(0, 1), (0, 2), (2, 3), (1,3)];
-        let raw_layout = vec![(0, (10.0, 0.0)), 
-                                                      (1, (5.0, 1.0)), 
-                                                      (2, (15.0, 1.0)), 
-                                                      (3, (10.0, 2.0))];
-
-        let parameters = PlotParameters {
-            label_width: 5,
-            scaling: 1,
-            aspect_ratio: 1.0
-        };
-        let scaled_layout = ScaledLayout::new(raw_layout, edges, &parameters, None);
-        info!("scaled_layout: {:#?}", scaled_layout.labels);
-
-
-        // This should only round and transpose the coordinates
-        let expected_labels = HashMap::from([
-            (0, ("●".to_string(), 0, 10)),
-            (1, ("●".to_string(), 1, 5)),
-            (2, ("●".to_string(), 1, 15)),
-            (3, ("●".to_string(), 2, 10)),
-        ]);
-        assert_eq!(scaled_layout.labels, expected_labels);
-
-        
-    }
-
-    #[test]
-    fn test_scaled_layout_new_unlabeled_scaled() {
-        let _ = env_logger::try_init();
-
-        let edges = vec![(0, 1), (0, 2), (2, 3), (1,3)];
-        let raw_layout = vec![(0, (10.0, 0.0)), 
-                                                      (1, (5.0, 1.0)), 
-                                                      (2, (15.0, 1.0)), 
-                                                      (3, (10.0, 2.0))];
-
-        let parameters = PlotParameters {
-            label_width: 5,
-            scaling: 10,
-            aspect_ratio: 1.0
-        };
-        let scaled_layout = ScaledLayout::new(raw_layout, edges, &parameters, None);
-        info!("scaled_layout: {:#?}", scaled_layout.labels);
-
-        let expected_labels = HashMap::from([
-            (0, ("●".to_string(), 0, 100)),
-            (1, ("●".to_string(), 10, 50)),
-            (2, ("●".to_string(), 10, 150)),
-            (3, ("●".to_string(), 20, 100)),
-        ]);
-        assert_eq!(scaled_layout.labels, expected_labels);
-    }
-
-    #[test]
-    fn test_scaled_layout_new() {
-        let _ = env_logger::try_init();
-
-        let edges = vec![(0, 1), (0, 2), (2, 3), (1,3)];
-        let raw_layout = vec![(0, (10.0, 0.0)),
-                                                      (1, (5.0, 1.0)), 
-                                                      (2, (15.0, 1.0)), 
-                                                      (3, (10.0, 2.0))];
-        let full_labels = HashMap::from([
-            (0, "ABCDEFGH".to_string()), 
-            (1, "IJKLMNOP".to_string()), 
-            (2, "QRSTUV".to_string()), 
-            (3, "WXYZ".to_string())
-        ]);
-
-        let parameters: PlotParameters = PlotParameters {
-            label_width: u32::MAX,
-            scaling: 1,
-            aspect_ratio: 1.0
-        };
-        let scaled_layout = ScaledLayout::new(raw_layout, edges, &parameters, Some(full_labels));
-        info!("scaled_layout: {:#?}", scaled_layout.labels);
-
-        let expected_labels = HashMap::from([
-            (0, ("ABCDEFGH".to_string(), 0, 10)),
-            (1, ("IJKLMNOP".to_string(), 9, 5)),
-            (2, ("QRSTUV".to_string(), 10, 15)),
-            (3, ("WXYZ".to_string(), 18, 10)),
-        ]);
-        assert_eq!(scaled_layout.labels, expected_labels);
-    }
-
-    #[test]
-    fn test_scaled_layout_new_truncations() {
-        let _ = env_logger::try_init();
-
-        let edges = vec![(0, 1), (0, 2), (2, 3), (1,3)];
-        let raw_layout = vec![(0, (10.0, 0.0)), (1, (5.0, 1.0)), (2, (15.0, 1.0)), (3, (10.0, 2.0))];
-        let full_labels = HashMap::from([
-            (0, "ABCDEFGH".to_string()), 
-            (1, "IJKLMNOP".to_string()), 
-            (2, "QRSTUV".to_string()), 
-            (3, "WXYZ".to_string())
-        ]);
-
-        let parameters = PlotParameters {
-            label_width: 5,
-            scaling: 1,
-            aspect_ratio: 1.0
-        };
-        let scaled_layout = ScaledLayout::new(raw_layout, edges, &parameters, Some(full_labels));
-        info!("scaled_layout: {:#?}", scaled_layout.labels);
-
-        let expected_labels = HashMap::from([
-            (0, ("A...H".to_string(), 0, 10)),
-            (1, ("I...P".to_string(), 6, 5)),
-            (2, ("Q...V".to_string(), 6, 15)),
-            (3, ("WXYZ".to_string(), 12, 10)),
-        ]);
-        assert_eq!(scaled_layout.labels, expected_labels);
-    }
-
-    #[test]
-    fn test_scaled_layout_new_edges() {
-        let _ = env_logger::try_init();
-
-        let edges = vec![(0, 1)];
-        let raw_layout = vec![(0, (5.0, 0.0)),
-                                                      (1, (5.0, 10.0))];
-        let full_labels = HashMap::from([
-            (0, "ABCDEFGH".to_string()), 
-            (1, "IJKLMNOP".to_string())
-        ]);
-
-        let parameters: PlotParameters = PlotParameters {
-            label_width: u32::MAX,
-            scaling: 1,
-            aspect_ratio: 1.0
-        };
-        let scaled_layout = ScaledLayout::new(raw_layout, edges, &parameters, Some(full_labels));
-
-        let expected_labels = HashMap::from([
-            (0, ("ABCDEFGH".to_string(), 0, 5)),
-            (1, ("IJKLMNOP".to_string(), 18, 5))]);
-        assert_eq!(scaled_layout.labels, expected_labels);
-
-        let expected_lines = vec![((8.0, 5.5), (16.5, 5.5))];
-        assert_eq!(scaled_layout.lines, expected_lines);
-    }
+ 
 }
